requirements.txt
tensorflow==2.13.0
opencv-python==4.8.0.74
numpy==1.24.3
flask==2.3.2
flask-cors==4.0.0
RPi.GPIO==0.7.1
gpiozero==1.6.2
psutil==5.9.5
requests==2.31.0
